Move `ExternalTable` and `ExternalPartition` to `source_definition` module in dbt/artifacts
