+++ core/dbt/artifacts/resources/__init__.py
@@ -58,3 +58,7 @@ from dbt.artifacts.resources.v1.semantic_model import (
     SemanticModel,
     SemanticModelConfig,
 )
+from dbt.artifacts.resources.v1.source_definition import (
+    ExternalPartition,
+    ExternalTable,
+)
