+++ core/dbt/artifacts/resources/v1/source_definition.py
@@ -1,0 +1,29 @@
+from dataclasses import dataclass, field
+from dbt_common.contracts.config.properties import AdditionalPropertiesAllowed
+from dbt_common.contracts.util import Mergeable, Replaceable
+from dbt_common.exceptions import CompilationError
+from typing import Any, Dict, List, Optional, Union
+
+
+@dataclass
+class ExternalPartition(AdditionalPropertiesAllowed, Replaceable):
+    name: str = ""
+    description: str = ""
+    data_type: str = ""
+    meta: Dict[str, Any] = field(default_factory=dict)
+
+    def __post_init__(self):
+        if self.name == "" or self.data_type == "":
+            raise CompilationError("External partition columns must have names and data types")
+
+
+@dataclass
+class ExternalTable(AdditionalPropertiesAllowed, Mergeable):
+    location: Optional[str] = None
+    file_format: Optional[str] = None
+    row_format: Optional[str] = None
+    tbl_properties: Optional[str] = None
+    partitions: Optional[Union[List[str], List[ExternalPartition]]] = None
+
+    def __bool__(self):
+        return self.location is not None
